requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
dateparser==0.7.2
decorator==4.4.1
Flask==1.1.1
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-automl==0.8.0
google-cloud-vision==0.40.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.0
idna==2.8
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.4
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
regex==2019.11.1
requests==2.22.0
rsa==4.0
scikit-image==0.16.2
scipy==1.3.2
six==1.13.0
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
